+++ src/HeboTech.ATLib/DTOs/PhoneNumberDTO.cs
@@ -1,5 +1,8 @@
 ﻿namespace HeboTech.ATLib.DTOs
 {
+    /// <summary>
+    /// Phone number received from the modem
+    /// </summary>
     public class PhoneNumberDTO
     {
         public PhoneNumberDTO(string number)
@@ -7,6 +10,9 @@
             Number = number;
         }
 
+        /// <summary>
+        /// The phone number received from the modem
+        /// </summary>
         public string Number { get; }
 
         public override string ToString()
