--- src/HeboTech.ATLib/PDU/IEI.cs
@@ -1,7 +1,0 @@
-﻿namespace HeboTech.ATLib.PDU
-{
-    internal enum IEI
-    {
-        ConcatenatedShortMessages8BitReferenceNumber = 0x00,
-    }
-}
